+++ Main Project/bossFight.html
@@ -10,11 +10,11 @@
 	<script src="libraries/p5.sound.min.js"></script>
 	<script src="gamepad.js"></script>
     <script  src="enemy.js"></script>
-    <script  src="player.js"></script>
-	<script src="gameObject.js"></script>
-	<script src="projectile.js"></script>
-	<script src="boss.js"></script>
-	<script src="bossFight.js"></script>
+    <script  src="./gameClasses/player.js"></script>
+	<script src="./gameClasses/gameObject.js"></script>
+	<script src="./gameClasses/projectile.js"></script>
+	<script src="./gameClasses/boss.js"></script>
+	<script src="./gameClasses/bossFight.js"></script>
 </head>
 <body>
     <div id="health-bar-container">
