+++ Main Project/gameClasses/upgradeIslandObjects.js
