+++ Main Project/index.html
@@ -12,10 +12,10 @@
   <script src="libraries/p5.sound.min.js"></script>
   <script src="libraries/p5.collide2D.js"></script>
   <script src="gamepad.js"></script>
-  <script src="player.js"></script>
-  <script src="gameObject.js"></script>
-  <script src="enemy.js"></script>
-  <script src="projectile.js"></script>
+  <script src="./gameClasses/player.js"></script>
+  <script src="./gameClasses/gameObject.js"></script>
+  <script src="./gameClasses/enemy.js"></script>
+  <script src="./gameClasses/projectile.js"></script>
 
 </head>
 
@@ -24,7 +24,7 @@
 		<div id="health-bar"></div>
 	</div>
 	<div class = "coins">Dabloons: <span id = "coinCount"></span></div>
-  <script src="sketch.js"></script>
+  <script src="./sketches/sketch.js"></script>
   <script>
 	// Initial update to display the starting currency
 	document.addEventListener('DOMContentLoaded', function() {
