+++ Main Project/islandIndex.html
@@ -53,11 +53,11 @@
     <script src="libraries/p5.min.js"></script>
     <script src="libraries/p5.sound.min.js"></script>
     <script src="libraries/p5.collide2D.js"></script>
-    <script src="playerPirate.js"></script>
-    <script src="islandObjects.js"></script>
+    <script src="./gameClasses/playerPirate.js"></script>
+    <script src="./gameClasses/islandObjects.js"></script>
 
 </head>
 <body>
-    <script src="sketchIsland.js"></script>
+    <script src="./sketches/sketchIsland.js"></script>
 </body>
 </html>
