+++ Main Project/nurikabe.html
@@ -13,13 +13,13 @@
     <div class="coins" style="position: absolute; top: 260px; left: 10px; z-index: 1000;">
         Coins: <span id="coinCount">0</span>
     </div>
-    <script src="player.js"></script> <!-- Ensure player.js is included -->
-		<script src="sketch.js"></script> <!-- Ensure sketch.js is included -->
+    <script src="./gameClasses/player.js"></script> <!-- Ensure ./gameClasses/player.js is included -->
+		<script src="./sketches.sketch.js"></script> <!-- Ensure ./sketches.sketch.js is included -->
 		<script>	
 			// Call the function to initialize the coin count display
 			player.updateCoinCount();
 		</script>
-		<script src="nurikabe.js"></script>
+		<script src="./puzzleFiles/nurikabe.js"></script>
     <button onclick="window.location.href='index.html'" style="position: absolute; top: 10px; left: 10px; z-index: 1000;">Back to Game</button>
   </body>
   <style>
