+++ Main Project/sudoku.html
@@ -32,8 +32,8 @@
     <button id="hintButton" onclick="giveHint()" style="position: absolute; top: 75%; left: 56%; transform: translate(-50%, -50%); z-index: 1000;">Hint</button>
     </button>
     <button id="finishButton" onclick="window.location.href='index.html'" style="position: absolute; top: 60%; left: 50%; transform: translate(-50%, -50%); z-index: 1000;">Collect Treasure</button>
-    <script src="sudokuPick.js"></script>
-    <script src="sudoku.js"></script>
+    <script src="./puzzleFiles/sudokuPick.js"></script>
+    <script src="./puzzleFiles/sudoku.js"></script>
   </body>
   <style>
     body{
