+++ Main Project/test/boss.test.js
@@ -1,6 +1,6 @@
-const  Boss  = require('../boss');
-const  Player  = require('../player');
-const  Projectile  = require('../projectile');
+const  Boss  = require('../gameClasses/boss');
+const  Player  = require('../gameClasses/player');
+const  Projectile  = require('../gameClasses/projectile');
 
 // Mock localStorage
 global.localStorage = {
