+++ Main Project/upgrade.html
@@ -57,12 +57,12 @@ canvas{
 <script src="libraries/p5.sound.min.js"></script>
 <script src="libraries/p5.collide2D.js"></script>
 <script src="gamepad.js"></script>
-<!-- <script src="player.js"></script> -->
-<!-- <script src="gameObject.js"></script> -->
-<!-- <script src="enemy.js"></script> -->
-<script src="projectile.js"></script>
+<!-- <script src="./gameClasses/player.js"></script> -->
+<!-- <script src="./gameClasses/gameObject.js"></script> -->
+<!-- <script src="./gameClasses/enemy.js"></script> -->
+<script src="./gameClasses/projectile.js"></script>
 <!-- <script src="islandObject.js"></script> -->
-<script src="pirate.js"></script>
+<script src="./gameClasses/pirate.js"></script>
 
 </head>
 <body>
